Move Jacobian matrix function into `dswx_sar_helper.py`
